JAMES-3008 Integration tests for CassandraBlobExistenceTester + swift
